ADDED: Mild refactor
